feat :Implement geolocation-based toplist selection using CF-IPCountry

+++ database/migrations/2025_06_14_131550_create_brands_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
            $table->id('brand_id');
         $table->string('brand_name');
         $table->string('brand_image');
-        $table->integer('rating');
+        $table->float('rating');
         });
     }
